Refactor test utils, redux store and hooks files
